lab4_2/src/python/requirements.txt
paho-mqtt>=1.6.1
matplotlib>=3.0

# Optional: for nicer plotting, install seaborn
seaborn>=0.11.0
